Final fix scanf sscanf printf, y blobWars anda perfecto salvo el tablero

+++ Kernel/naiveConsole.c
@@ -26,7 +26,7 @@ void ncPrintChar(char character)
 		ncNewline();
 	}
 	else{
-		*currentVideo = character;
+		*currentVideo = (uint8_t)character;
 		currentVideo += 2;
 	}
 }
